Refactor: Rename plugin package to grpcplugin
