+++ backend/grpcplugin/grpc_data.go
@@ -1,4 +1,4 @@
-package plugin
+package grpcplugin
 
 import (
 	"context"
